Move FileSizeInput component to core
