+++ frontend/src/components/core/FileSizeInput.tsx
